requirements.txt
python-telegram-bot==22.3
python-dotenv==1.1.1
gspread==6.2.1
flask==2.1.0  # Only if you use `keep_alive.py` with Flask